Move logic for #crack into its own class
